Refactored project. ezmvp package pending removal.
