+++ src/main/java/com/nestorledon/ezapp/ActivityBase.java
@@ -1,6 +1,5 @@
-package com.nestorledon.ezapp.base;
+package com.nestorledon.ezapp;
 
-import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
@@ -10,18 +9,15 @@ import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.View;
-import android.widget.Adapter;
 import android.widget.AdapterView;
 import android.widget.ListAdapter;
 import android.widget.ListView;
 
-import com.nestorledon.ezapp.R;
+import com.nestorledon.ezapp.base.Navigator;
+import com.nestorledon.ezapp.base.PagerAdapter;
 import com.nestorledon.ezapp.base.widgets.AdjustedDrawerLayout;
-import com.nestorledon.ezapp.base.widgets.SlidingTabLayout;
+import com.nestorledon.ezapp.base.widgets.slidingtab.SlidingTabLayout;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 
 /**
