+++ src/main/java/com/nestorledon/ezmvp/Informer.java
@@ -1,4 +1,4 @@
-package com.nestorledon.ezapp.base;
+package com.nestorledon.ezmvp;
 
 /**
  * Created by nestorledon on 4/30/15.
