+++ src/main/java/com/nestorledon/ezmvp/event/EventArgs.java
@@ -1,4 +1,4 @@
-package com.nestorledon.ezapp.base.event;
+package com.nestorledon.ezmvp.event;
 
 import java.util.HashMap;
