+++ src/main/java/com/nestorledon/ezmvp/forms/HasLong.java
@@ -1,4 +1,4 @@
-package com.nestorledon.ezapp.base.forms;
+package com.nestorledon.ezmvp.forms;
 
 /**
  * Created by nestorledon on 5/2/15.
